+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Implement `Iter.fallibleForEach` and `Iter.fallibleFold`
+- Implement `Iter.fallibleReduce` as private, experimental methods
+
+### Changed
+
+- Make `Iter.sizeHint` public but still experimental
+- Removed `Take.curr` field
+
+### Fixed
+
+- Fixed tests names
+- Fixed `reiter.recurse` example in README 
+
 ## [0.3.0] - 2024-05-20
 
 ### Added